refactor(app): isolate delivery wiring
